final_project/backend/requirements.txt
flask
sqlalchemy
flask_sqlalchemy
sqlalchemy-utils
marshmallow
mysql-python
oauth2client
requests
pytest
coverage
pyjwt